cbserializer/requirements.txt
Django>4.0<=4.0.6
djangorestframework>3.13.0<=3.13.1
pymongo==3.12.3
# snappy>3.0<=3.0.3
# gssapi>1.7.0<=1.8.1
srv>1.0.2<=1.1.0
dnspython>1.14.0<=1.15.0
djongo>1.3.4<=1.3.6
pygments>2.11.0<=2.13.0